mips: fix compilation warning
* syscall.c (get_scno) [MIPS]: Fix harmless compilation warning
about comparison between signed and unsigned integer expressions.

+++ syscall.c
@@ -1407,7 +1407,7 @@ get_scno(struct tcb *tcp)
 	scno = mips_REG_V0;
 
 	if (!SCNO_IN_RANGE(scno)) {
-		if (mips_REG_A3 == 0 || mips_REG_A3 == -1) {
+		if (mips_REG_A3 == 0 || mips_REG_A3 == (uint64_t) -1) {
 			if (debug_flag)
 				fprintf(stderr, "stray syscall exit: v0 = %ld\n", scno);
 			return 0;
